requirements.txt
altair==5.4.0
annotated-types==0.7.0
asgiref==3.7.2
attrs==24.2.0

blis==1.0.1
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.0
cymem==2.0.10
distlib==0.3.8
Django==5.0.3
dnspython==2.7.0
docstring_parser==0.16
faiss-cpu==1.9.0.post1
filelock==3.13.1

fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.73.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.62.3
httplib2==0.22.0
huggingface-hub==0.26.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.11
narwhals==1.3.0
networkx==3.4.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pip==24.3.1
platformdirs==4.2.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==65.5.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.11.0
sqlparse==0.4.4
srsly==2.4.8
streamlit==1.37.1
streamlit-lottie==0.0.5
sympy==1.13.1
tenacity==8.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.46.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.25.1
wasabi==1.1.3
watchdog==4.0.2
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.0
soundfile==0.12.1
pydub==0.25.1
sounddevice==0.4.6  # Added for microphone recording support
PyPDF2==2.7.0
